Make transcript selection reproducible
In transcript selection mode, the selected transcript may depend on the
order the transcripts are returned by the dict.items() method. This
order is not guaranteed to be predictable. We sort the list of candidate
transcripts by ENST. In this way we know, at least, that in case of ties
between candidates we consistently select the last one.

(See also comments in the script)